Organize formatters
