+++ lib/util.js
@@ -370,32 +370,6 @@ export async function preprocessData(data) {
   return data
 }
 
-export function formatCurrency(number) {
-  return `$${Number(number).toLocaleString()}`
-}
-
-export function formatCurrencyMillions(number) {
-  return `$${Math.floor(number / 1000000)}`
-}
-
-export function formatCurrencyWithUnit(number) {
-  if (number === 0) {
-    return '$0'
-  } else if (number < 1000000) {
-    return `$${Math.floor(number / 1000)}k`
-  } else {
-    return `$${Math.floor(number / 1000000)}m`
-  }
-}
-
-export function formatCategory(item) {
-  if (!item.fields['Parent Category'].id || item.fields.Category.id === item.fields['Parent Category'].id) {
-    return item.fields.Category.fields.Name
-  } else {
-    return item.fields['Parent Category'].fields.Name
-  }
-}
-
 export function isValidBbox(bbox) {
   if (bbox[0] > 180 || bbox[0] < -180) {
     return false
